--- src/standalone_custom/heat_requirements.txt
@@ -1,1 +1,0 @@
-oemof.thermal==0.0.2
